# backend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LumiCreate-Pro API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/subtitle-engine/status/{project_id}": {
      "get": {
        "tags": [
          "subtitle-engine"
        ],
        "summary": "Subtitle Status",
        "operationId": "subtitle_status_api_subtitle_engine_status__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubtitleStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subtitle-engine/script/{project_id}": {
      "get": {
        "tags": [
          "subtitle-engine"
        ],
        "summary": "Get Script",
        "description": "Return subtitle script lines extracted from the project's scenes (dialogue / description).",
        "operationId": "get_script_api_subtitle_engine_script__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subtitle-engine/preprocess-text": {
      "post": {
        "tags": [
          "subtitle-engine"
        ],
        "summary": "Preprocess Text Endpoint",
        "operationId": "preprocess_text_endpoint_api_subtitle_engine_preprocess_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreprocessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subtitle-engine/generate-srt": {
      "post": {
        "tags": [
          "subtitle-engine"
        ],
        "summary": "Generate Srt Endpoint",
        "description": "SSE stream: normalize video → extract audio → align → write SRT.",
        "operationId": "generate_srt_endpoint_api_subtitle_engine_generate_srt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSrtRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subtitle-engine/embed": {
      "post": {
        "tags": [
          "subtitle-engine"
        ],
        "summary": "Embed Subtitles Endpoint",
        "description": "Burn subtitles.srt into final_video.mp4 → final_video_subbed.mp4.",
        "operationId": "embed_subtitles_endpoint_api_subtitle_engine_embed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmbedRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "font_name": {
            "type": "string",
            "title": "Font Name",
            "default": "等线 Bold"
          },
          "font_size": {
            "type": "integer",
            "title": "Font Size",
            "default": 18
          },
          "primary_colour": {
            "type": "string",
            "title": "Primary Colour",
            "default": "#FFFFFF"
          },
          "outline_colour": {
            "type": "string",
            "title": "Outline Colour",
            "default": "#000000"
          },
          "outline_width": {
            "type": "number",
            "title": "Outline Width",
            "default": 2.0
          },
          "shadow_depth": {
            "type": "number",
            "title": "Shadow Depth",
            "default": 0.0
          },
          "margin_v": {
            "type": "integer",
            "title": "Margin V",
            "default": 30
          },
          "position": {
            "type": "string",
            "title": "Position",
            "default": "bottom"
          },
          "bold": {
            "type": "boolean",
            "title": "Bold",
            "default": true
          },
          "italic": {
            "type": "boolean",
            "title": "Italic",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "EmbedRequest"
      },
...
      "GenerateSrtRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "lines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Lines"
          },
          "fps": {
            "type": "integer",
            "title": "Fps",
            "default": 24
          },
          "manual_advance": {
            "type": "number",
            "title": "Manual Advance",
            "default": 0.0
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "default": "base"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "lines"
        ],
        "title": "GenerateSrtRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreprocessRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "PreprocessRequest"
      },
...
      "SubtitleStatus": {
        "properties": {
          "has_final_video": {
            "type": "boolean",
            "title": "Has Final Video"
          },
          "has_fixed_cfr": {
            "type": "boolean",
            "title": "Has Fixed Cfr"
          },
          "has_srt": {
            "type": "boolean",
            "title": "Has Srt"
          },
          "has_embedded": {
            "type": "boolean",
            "title": "Has Embedded"
          },
          "srt_path": {
            "type": "string",
            "title": "Srt Path",
            "default": ""
          },
          "embedded_path": {
            "type": "string",
            "title": "Embedded Path",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "has_final_video",
          "has_fixed_cfr",
          "has_srt",
          "has_embedded"
        ],
        "title": "SubtitleStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}